requirements.txt
google-api-python-client
beautifulsoup4
pinecone
openai
requests
tqdm
